+++ components/Form/index.js
@@ -1,73 +1,20 @@
-import { useState } from "react";
 import InputFields from "../InputFields";
 import Checkbox from "../Checkbox";
 import SelectInput from "../SelectInput";
 
-export default function Form() {
-  const [age, setAge] = useState("");
-  const [height, setHeight] = useState("");
-  const [weight, setWeight] = useState("");
-  const [intensity, setIntensity] = useState("");
-  const [category, setCategory] = useState("All");
-
-  const [physique, setPhysique] = useState([
-    { name: "Slim", checked: false },
-    { name: "Normal", checked: false },
-    { name: "Strong", checked: false },
-  ]);
-  const [fitnessLevel, setFitnessLevel] = useState([
-    { name: "Unatheltic", checked: false },
-    { name: "Fit", checked: false },
-    { name: "Muscular", checked: false },
-  ]);
-  const [timePerWeek, setTimePerWeek] = useState([
-    { name: "2-3x", checked: false },
-    { name: "3-4x", checked: false },
-    { name: "4+", checked: false },
-  ]);
-  const [preference, setPreference] = useState([
-    { name: "Teamsport", checked: false },
-    { name: "Individual sports", checked: false },
-    { name: "Indoor", checked: false },
-    { name: "Outdoor", checked: false },
-  ]);
-
-  function handleAgeChange(event) {
-    setAge(event.target.value);
-  }
-
-  function handleHeightChange(event) {
-    setHeight(event.target.value);
-  }
-
-  function handleWeightChange(event) {
-    setWeight(event.target.value);
-  }
-
-  function handleIntensityChange(event) {
-    setIntensity(event.target.value);
-  }
-
-  function handleCategoryChange(event) {
-    setCategory(event.target.value);
-  }
-
-  function handleChange(type, name) {
-    const stateVariables = {
-      physique: [physique, setPhysique],
-      fitnessLevel: [fitnessLevel, setFitnessLevel],
-      timePerWeek: [timePerWeek, setTimePerWeek],
-      preference: [preference, setPreference],
-    };
-
-    const [state, setState] = stateVariables[type];
-    const newState = state.map((item) =>
-      item.name === name ? { ...item, checked: !item.checked } : item
-    );
-
-    setState(newState);
-  }
-
+export default function Form({
+  weight,
+  height,
+  age,
+  intensity,
+  physique,
+  fitnessLevel,
+  timePerWeek,
+  handleChange,
+  onChange,
+  category,
+  preference,
+}) {
   return (
     <form aria-labelledby="user-details">
       <legend>
@@ -78,21 +25,21 @@ export default function Form() {
         type="text"
         name="age"
         value={age}
-        onChange={handleAgeChange}
+        onChange={onChange}
       />
       <InputFields
         label="Weight(kg):"
         type="text"
         name="weight"
         value={weight}
-        onChange={handleWeightChange}
+        onChange={onChange}
       />
       <InputFields
         label="Height(m):"
         type="text"
         name="height"
         value={height}
-        onChange={handleHeightChange}
+        onChange={onChange}
       />
 
       <h2>Physique</h2>
@@ -135,8 +82,9 @@ export default function Form() {
         type="range"
         min={20}
         max={90}
+        name="intensity"
         value={intensity}
-        onChange={handleIntensityChange}
+        onChange={onChange}
       />
       <div>{intensity}%</div>
 
@@ -151,7 +99,7 @@ export default function Form() {
         />
       ))}
 
-      <SelectInput onChange={handleCategoryChange} />
+      <SelectInput onChange={onChange} />
       <p>Selected category: {category}</p>
 
       <button type="submit">Synergy</button>
